Using nested routes for integrations
