+++ client/pages/integrations/home/connect-button.js
