+++ client/pages/integrations/home/icon.js
