+++ client/pages/integrations/home/search-icon.js
