+++ client/pages/integrations/integration/index.js
@@ -1,0 +1,17 @@
+import { getIntegrationModule } from '@quillforms/form-integrations';
+
+const Integration = ( { slug, isLoading } ) => {
+	const integrationModule = getIntegrationModule( slug );
+
+	return (
+		<div className="quillforms-single-integration">
+			{ ! isLoading ? (
+				<integrationModule.render />
+			) : (
+				'Loading integration page...'
+			) }
+		</div>
+	);
+};
+
+export default Integration;
