+++ packages/navigation/src/index.ts
@@ -8,6 +8,7 @@ export {
 	matchPath,
 	BrowserRouter,
 	useParams,
+	useRouteMatch,
 	Redirect,
 } from 'react-router-dom';
 export { default as NavigationPrompt } from 'react-router-navigation-prompt';
